requirements.txt
numpy 
pandas
scikit-learn
matplotlib
seaborn
tensorflow
keras
torch
google-cloud-storage
pyyaml
imbalanced-learn
joblib
xgboost
lightgbm
mlflow
flask